# swagger/yaml/openapi.yaml
openapi: 3.0.0
info:
  title: Product
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
paths:
...
  /registration:
    post:
      tags:
        - default
      summary: Registration
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: Ghaida Fasya Yuthika Afifah
                npm: 714220031
                numberPhone: '081243243133'
                password: admin12345
                confirmPassword: admin12345
                role: Buyer
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /login:
    post:
      tags:
        - default
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                npm: 714220032
                password: admin12345
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /category:
    get:
      tags:
        - default
      summary: AllCategory
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - default
      summary: AddCategorySeller
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                id_category_admin: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /category-seller:
    get:
      tags:
        - default
      summary: CategorySeller
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /category-admin:
    get:
      tags:
        - default
      summary: AllCategoryAdmin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /category-admin/:
    post:
      tags:
        - default
      summary: AddCategoryAdmin
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                category: Gaming
                icon: fa-solid-fa-mobile
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...